* Improve Config API for most common usages
This patch introduces a few more (and hopefully the last!) API changes to
the ConfigParser class.

The get() method is changed to:

  T get(char[] category, char[] key, T default_value);

So a common use case like this:

  bool activate_something = false;
  Config.get(activate_something, cat, key);
  if (activate_something)
  {
      do_something();
  }

Can be reduced to:

  if (Config.get(cat, key, false))
  {
      do_something();
  }

If you need to know if a configuration variable was actually set, you
can always use the Config.exists() or Config.getStrict() methods.

Config.getList() was also updated to this kind of API, and also the
template parameter specification was changed, so it can now get no only
lists of strings, but lists of arbitrary atomic types (at least the ones
supported by get()). Both getList() and getListStrict() was fixed to
work with the latests Config changes.

Finally the Config.isRead() method was renamed to Config.isEmpty(),
because the isRead() name was misleading. The method only checks if some
values are present in the config object, but those values can be set in
other ways besides reading a config file, or even an empty config file
could have been read.

The documentation (specially the examples) are also updated.