+ Add a new framework to easily create applications
A new package is added, ocean.util.app, which implements a framework to
easily create robust and consistent applications, saving most of the setup
work. The framework is extremely flexible and extensible.

Creating an application is as simple as deriving from one of the provided
applications (or create a new type) and overriding the run() method.
Overriding other methods you can add more command line arguments, parse
config files, etc.

See the documentation in the code for more details and examples, specially
the documentation for ocean.util.app.Application and ocean.util.app.*App.