* Fix unknown struct size compile error
The size of the struct can be known in the declaration of the struct
itself, specially when it's potentialy stored in a struct member. This
patch convert the length member to a static function (which should be
inlined anyway, so there should be no performance loss).

I also think the SizeofTuple template should be removed in favor of
sizeof, that template is only checking if the compiler attribute
align(1) works. If we can't trust the compiler for one thing (align(1)
and sizeof), I can't see how we can trust it in the other (the template)
:)