mkversion: Support mixed svn/git svn repos
The main repository and the libraries repositories now can be an arbitrary
mix of svn and git svn repositories.

This is mainly needed because Codespaces has upgraded to subversion 1.7
which introduces some backwards incompatible changes that makes git svn
fail on some repos (like swarm) because of invalid tag names. Now the
problematic repos can be used with svn directly while keeping all the other
repos in git svn.