* Derive Fnv1 classes from FnvDigest super-class
All Fnv classes are instances of the FnvGeneric template, but they lack
a common interface besides Tango's Digest, which is a little too generic
and hard to use for most cases.

This patch introduces a new super-class to all theses classes,
FnvDigest, which for now only have the ulongDigest() method. The reset()
method is a good candidate to be part of the super-class but an API
change is needed to do that because now it returns an instance of the
concrete class and it should be changed to return an instance of the
FnvDigest class instead.