* Add usage and help strings to text.Arguments
This patch adds text.Arguments the capability of printing more information
when asking for help. A short description, usage string and a detailed
string are included (but optional). Also the application name is stored.
All the new strings can be set in the constructor and displayHelp() taking
the app_name argument is deprecated.

The util.app framework is updated to take advantage of this, so some other
methods constructors are deprecated in the affected classes.