+ Fix infinite loop in StringEncoding.convert
If input was the single char ä, then it doesn't fit in the output buffer. It catches the exception
and then tries again. But, before the exception is thrown, the output buffer gets shrunk
to the part that was used. This takes it back to the original size...  Infinite loop.
This caused shore-feed to hang.