FlexibleRingQueue: Fix how willFit() is calculated
FlexibleRingQueue.willFit() doesn't take into account where the read
pointer is, so when it's at the end, and the write pointer is at the end
too, willFit() can return true even when the queue is really full.

Also add a test to the unittest so it won't happen again.

Fixes #5.