Add new AppStatus display log.
The AppStatus display log consists of a static section at the bottom of the terminal window and a streaming section above that. The static section
includes infromation such as the current runtime, memory usage, cpu usage, and version information of the application.

Other changes include a new MessageOnlyLayout that only prints the message of the log event and no other information and two extra formatting
methods to Stdout (up and endrow).