Add option to open existing file queues
When saving existing file queues, will write an index file in addition
to the data file. The index file is written whenever information is
read or written to the data file.

To support the reading of existing file queues on start up there is now
a new enablePop function. It defaults to false and the file queue will
not pop any items. When set to true the queue will pop again. The
calling application will need to call the enablePop method once the
connection to the dht has been established.