Add StackAllocated template class, wrapping a StackAllocator and an object
The typical usage pattern of the StackAllocator is to new one at scope,
then at some point later to use it to allocate an instance of a class
derived from a known base. The new StackAllocated class essentially wraps
a StackAllocator with a class instance, making it easier to pass these
things around between methods and so on.