mkversion: Use (tag|branch)-hash[!] as git version
Since we are not making real use of tags, using the normal "git describe"
format doesn't make a lot of sense. Instead of that, we manually construct
a description of a revision that suits us.

What we do is to build our version using a symbolic name and a hash always.
If a tag matches exactly the current HEAD, the tag name is used as the
symbolic name. Otherwise the branch name is used (or the string "DETACHED"
if we are using a "detached HEAD". The current hash is always included
(even when using the tag name, as the tag could be changed or removed).
Finally, the optional "!" at the end indicated a dirty repository (meaning
there are uncommitted changes in the repository at the time of compiling).