Map: Reference of key should be actual reference
When iterating over a map using ref for the key, the reference is not an actual
reference to the key but instead a reference to a local variable. This is very
unexpected and dangerous as bugs caused by this can cause bad memory
corruption. In my case it was just a segfault, but still.

It was supposedly done, to check with the assert whether the key has been
changed. With this commit, this check still works, but the actual key is passed
as reference.