Fix issue  #93  Incorrect URL percent decoding
The existing code was completely wrong, but unfortunately the front-end
has been relying on this wrong behaviour.

To retain backwards compatibility, first check to see if the
percent encoding could possibly be UTF8. If it isn't, fall back to
the old, incorrect method. This detection isn't quite perfect, it
may fail in the case where a character in the range C0-DF is followed
by one in the range 80-BF. Most of the ones in the 80-BF range are
wierd control characters, except non-breaking space and «¡©®£¢¥.
and the ones in range C0-DF are accented upper case A, E, I, O, U, plus
ß, Ø, and Ð. In the case of something like "CDÄ®", the Ä® would be
mistakenly processed as a single UTF8 character.
This situation is likely to be extremely rare since the
currency symbols would almost always be preceded by a (non-breaking) space.

It will never fail for percent encoding which follows the standard.