Preallocate cache mapping elements in dynamic arrays
The cache uses internally an ebtree and a hash map which are in general
designed to hold any number of elements and therefore allocate each element
(ebtree node or bucket element, respectively) separately using `new`.
However, for usage inside the cache the elements can be preallocated in
dynamic array buffers since since the maximum number of elements is known
in advance and all elements will be used.
The `ArrayPool` class implements this, using `malloc()` rather than the D
memory manager to avoid unnecessary spare allocation and make the elements
invisible to the GC.
Making the EBTree extensible to use a custom allocation method obsolesces
`minimize()` because it deletes all nodes, one by one, which isn't possible
if they are actually array elements. I chose to brutally remove
`minimize()` because it doesn't seem to be used anywhere.
A caveat is the requirement of the libebtree that the node pointers must be
integer multiples of 16. This holds true if each node instance is allocated
using `malloc()` or `new Node` but not if they are pointers to array
elements. It is now ensured by pad bytes following each node array element
and a postcondition of `NodePool.newNode()`.