Fix overflow detection logic
The previous logic for detecting integer overflows was somewhat naive, and
did not work in all cases.

This is especially clear when one considers the hexadecimal case. The old
logic was:

auto old_value = value;
value = value * radix + c;
if ( value < old_value ) // integer overflow
{
    return false;
}

Consider the conversion of the number 0x123456789. When reaching the 9
digit, `old_value` would be 0x12345678. `value * radix` is the equivalent
of a left shift by 4 bits (*16), which results in the number 0x23456780.
This number is larger than `old_value`, causing the overflow detection
logic to fail.