Refactor CompiledRegex class so it can be used outside of scope
The class was previously designed with the intention of it purely being
used as a scope class, thus:

auto pcre = new PCRE;
scope regex = pcre.compile("");
regex.match("");

However, when implementing https://github.com/sociomantic/swarm/issues/144,
it became apparent that this scope-restricted usage is not always possible.

This commit makes it possible to create an instance of CompiledRegex, and
to explicitly reuse it (to reinit it with a different regex string) without
destroying it.