StructDumper: only reset structs that contain dynamic arrays
When reseting all dynamic arrays in a struct to null only recurse in to
a sub struct field type if it contains dynamic arrays. Required for the
updates to the campaign metadata struct. Fixes a bug where the struct
dumper would fail a static assert when trying to dump a struct with
a dynamic array of sub structs and the sub struct has one field that is
a dymamic array of structs and another field that is also a struct (but
not a dynamic array).