Add new UTF-8 string truncate and append methods
The first method limits the length of a string by breaking at the
last whitespace before the limit.

The second method uses the first method to limit the length of a
string and also appends an ending to the string.

These methods were originally a single method in propulsor but have
been seperated and moved to ocean as they are needed in other projects.