Add DirectIO module to perform direct I/O
This kind of I/O is specially (and probably only) useful when you need
to dump a huge amount of data to disk as some kind of backup, i.e., data
that you'll probably won't ever read again (or at least in the short
term). In those cases, going through the page cache is not only probably
slower (because you are doing an extra memory copy), but it can
potentially freeze the whole system, if the sysctl vm.dirty_ratio is
passed, in that, programs will use their own time to write the page
cache to disk.  If the data you need to write is small or you are going
access it in the short term and you are not experiencing freezes, you
probably DON'T want to use this module.