Remove minilzo sources (and binaries)
To do this we rely on distribution packages (liblzo2-dev) to be
installed, and we switch to use the full lzo library instead of minilzo
(the API and ABI are the same so there should be no problems).

A `pragma(lib)` is added too so users don't need to care about the extra
linking flags.

Only code using the bare C bindings care about the module name change
and the symbols redistribution (according to liblzo2 include files
organization). Code using the D higher level bindings can remain
untouched.

This fixes #239.