Swap cache entries on removal
If a cache entry is removed, swap it and the last entry rather than copying
the last entry to it. Doing so avoids the situation that, if object
references (slices, objects or pointers) are stored in the cache and a new
cache element is created after removing elements, the value of the created
element references an object that is also referenced by another value in
the cache.