Override the dispose method of the AppStatus class
An application uses the AppStatus class when its console output needs to
be split into a streaming portion at the top, and a static portion at
the bottom. When such an application exits, the command prompt reappears
just above the static lines portion. This forces the user to press
'Enter' several times simply to skip over the static lines.

This commit overrides the dispose method of the AppStatus class in order
to insert a newline after the static lines portion. This way, when an
application using AppStatus exits, the command prompt appears
immediately after the static lines. Another benefit of using the dispose
method to insert a newline is that an application using AppStatus can,
if necessary, delete its AppStatus object (i.e. stop using the split
display mode) and output other content after the static lines at the
bottom.

Note that inserting the newline is done in the dispose method and not in
the destructor of the class to avoid the risk of the Stdout object
having been collected by the GC by the time the destructor executes.