Makd: Use `Version` as src/Version.d module name
At some point common.mk was updated to use the module name as `Version`
but Makd wasn't, so this inconsistency between both makes migration
harder.

Also, now `VERSION_MODULE` is not overrideable anymore, the module name
was computed based on the location and with the fixed module name it
makes no sense to do that, so overriding the location of the Version
module would be very confusing.