Add function to convert from a struct to a similar but not equal struct
This module allows you to convert a struct A to a similar but not equal
struct B. You can guide the conversion using converter functions for variables
that differ between them.  This module is mainly used for the soon to come
struct version loader where it will automatically convert from old versions to
the current one.