Makd: Check for validity of the used flavor
Specifying a flavor through the command-line can be error-prone, as
a simple typo could create a whole new flavor without you even noticing.
This commit check for the validity of the flavor. To check this, the
variable `$(VALID_FLAVORS)` was added, defaulting to `devel production`
(but it can be overridden in the `Config.mak` file to define more
-verified- flavors).

Fixes #320.