Add support for specifying layouts of logs
Currently a class inheriting from `LoggedCliApp.d`,
`VersionedLoggedCliApp.d` or `VersionedLoggedStatsCliApp.d`
automatically gets the capability of logging to the console and/or a log
file. Moreover some useful logging parameters can be automatically read
from a config file. However the layout of the log messages cannot be
configured by the user class (it always defaults to `LayoutDate` for
file logs and `SimpleLayout` for console logs).

This commit provides the ability for the user class to specify the
layout (separately for file logs and console logs).
The README is also updated with the log layout choice feature.

This commit would form the basis for a fix for sociomantic/shore#117