Fix warning in StructLoader when used with void[] arrays
IsPrimitive() is used to avoid instantinating & calling unneeded template
functions that wouldn't do anything with the given type anyway.
Those template functions however output a warning when they do nothing which can
be confusing and which also implies that IsPrimitive() wasn't checking all the
types it should have been checking, namely 'void'.

This commit adds the check for void to avoid this warning and unrequired calls
to mentioned template functions.