Don't be too flexible about layout string
The current implementation is too flexible with regard to the log layout
string given via the config file. For instance, a layout string as
absurd as `LAYOUTdLAYOUTaLAYOUTtLAYOUTeLAYOUT` passed to the
`file_layout` or `console_layout` keys would be accepted and treated as
`date`.

This is because currently all occurrences of `layout` within the given
layout string are removed prior to determining the layout to use.
Instead it is better to remove `layout` only from the beginning and/or
end of the layout string. This provides some level of flexibility
without accepting nonsense.