FlexibleFileQueue: Make unittest non-flakey
The unittest was doing I/O in a fixed file, which can fail and also
prevents different modules unittests to run in parallel (when the
involved modules includes this one) because all those tests operate over
the same test file.

This flakey test is fixed by creating a temporary directory for the test
using mkdtemp(), so each run of the unittest will use a different
directory. If the directory can't be created for any reason, then the
test is just skipped with an error message.