README: Switch Migration Instructions and New Features
It makes more sense to have the Migration Instructions first (this is
how it was historically) as it is more important for users to know how
to update their code than how to use new features.

This commit also removes a couple of trailing whitespaces and the
formatting of one item.