Restore template params used to specify log layout
Earlier commits removed the template parameters used to specify file and
console log layouts, since the layouts could be specified via the config
file itself.

However, there may be applications that want to define custom layouts
that are too application specific and not at all relevant to other
applications. Such non-generic layouts are not suitable to be added to
ocean, and are hence best supplied via the template parameters.