Do not suspend fiber or register timer event if s == 0
According to the man page for timerfd_settime:
  new_value.it_value specifies the initial expiration of the timer, in
  seconds and nanoseconds.  Setting either field of new_value.it_value
  to a nonzero value arms the timer.  Setting both fields of
  new_value.it_value to zero disarms the timer.

Previously, it was possible to set the timer to 0 in this way, causing it
to be disabled (i.e. never fire). Thus it was possible to get into a state
where the fiber was suspended waiting for an event which will never occur.

Now, if the timer would be disabled, we simply do nothing.