Restructure code to avoid extra indentation
In the `parseLine` function, `if` checks are used to ignore blank lines
and comment lines. Structuring code this way creates additional levels
of indentation and pushes code to the right. This can be easily avoided
by reversing the `if` checks and returning early from the function if
the check succeeds.