Introduce process monitor
Provide the ability for an application to use different
EpollSelectDispatcher instances with different EpollProcess instances.

Without this ability, it is currently not possible for an application to
create EpollProcess instances in a unittest block. This is because in
order to create an EpollProcess instance, an EpollSelectDispatcher
instance needs to be created first, and there is no way to get the
EpollSelectDispatcher instance created in a unittest block once the
application's 'main' function has been entered.

Adding this ability has the effect that the `RunningProcesses` class
gets renamed to `ProcessMonitor`; since this better reflects the nature
of the work done in this class.

Also add unit tests to confirm correct working of this feature.