StructConverter: Pick correct convertion function
The previous check only checked typeof(to.conver_*) which can give not the type
we check for if there exist several overloads of that function. By assigning it
to exactly the type we check for we can make sure to always check for the right
overload.

Add unit tests that verify such multiple signature conversion.