MapSerializer: Fix support for serialization of dynamic arrays
Previously, when loading maps with dynamic arrays as values, every value would
reference the same memory. Now the new array is copy'd into a potentially
existing array or a newly one is allocated.

It is not possible to add a unittest for this fix as it is linked to writing to
an actual file. It was however tested with one of my applications.