AppStatus: handle change in number of static lines
There are two issues with the existing implementation:
    1. If the number of static lines is increased, one or more lines in
       the upper streaming portion are overwritten and are thus lost.
    2. If the number of static lines is decreased, one or more of the
       previous static lines (including the static lines header) are
       "leaked" into the upper streaming portion.

This commit addresses both the above problems, and in the process
exposes a shortcoming in allowing the number of static lines to change
dynamically.
When the number of static lines is decreased, one or more blank lines
are created in the upper streaming portion. But this is to be expected
as a reduction in the static portion provides more room in the streaming
portion without any content to put in there.

Note that with the changes in this commit, there is no loss of
information in the upper streaming portion when the static portion is
increased or decreased.