Remove pragma(lib) for ebtree and lzo
Having automatic linking to libraries looks good on paper but sometimes
is not flexible enough because you can't control the order of the
libraries or how to do the linking (static vs. dynamic).

To make test pass now, we also need to add the appropriate link flags to
all ocean.io.compress.lzo modules now.