Use new ebtree modules location when specifying test flags
The ebtree modules were moved some time ago to a new location, but the
Build.mak file was never updated. The test kept passing because the
modules specified a `pragma(lib)`. In the next commits this pragma will
be removed so we need to update this locations before.

Also, to simplify the maintenance, a wildcard is used when appropriate
(any module in `ocean.util.container.ebtree` will most likely need to be
linked to ebtree anyway).