Makd: Replace unittest target with a test program
The current `make unittest` method is nice because you get a granular
report of which modules failed (if you use `-k`) but it's **extremely**
slow. Not only multiple files get compiled multiple times, the tests on
those files are also ran multiple times, making this scheme explode with
the number of files and inter-dependencies (it can potentially be
O(F²*T), where F is the number of files and T is the (average?) number of
test in each file).

This new method uses a new UnitTestRunner module to take care of how the
test are run, and consist on only generating a main file importing all
the modules in the project (plus this UnitTestRunner module).

The result is a single compilation of all the projects modules and one
single run for each unittest, greatly reducing the time to finish the
`make test` run.

For ocean, this means reducing the time to do `make clean; make test`
from 1 minute and 40 seconds (100 seconds) to 4.6 seconds (22x speedup).
For other projects it goes from 15x up to 40x speedup.

The new test runner also allows verbose output and keep going if
a module's unittest fails, options that are automatically used if `V=1`
or `-k` (respectively) are used when running make.