util.container.map.model.Bucket: Move val to the top
This is needed to guarantee the proper alignment of the value data, so
it can be scanned by the GC properly. For some reason, ubyte static
arrays get packed in the struct, so if the previous field was any type
not multiple of size_t.sizeof, this data will not be aligned to word
boundaries.

Fixes #369.