time.UnixTime: Enable and fix unittest
The test is broken because it relies on the local timezone, which you
can't really know. For now the tests using the local time are just
disabled. They could be fixed by querying the local timezone but this
module is deprecated anyway, so it doesn't seem worthy.

To enable the tests, OceanUnitTest was replaced with UnitTest.