core.ObjectPool: Enable and fix unittests
For some reason, when the ObjectPool to be tested is in the stack (scope
pool = ...), some other tests break, stuck in an infinite loop in
Tango's buffer handling. All points to stack corruption. Since this
module is deprecated, I don't think it worth investigating too much, so
for now the test is fixed by allocating in the GC heap instead.

To enable the unittest, OceanUnitTest was replaced by UnitTest.