UnitTestRunner: Add multiple --verbose levels
Now `--verbose` have 3 levels of verboseness, which can be specified by
adding multiple `--verbose` or `-v` arguments (or the compact for `-vv`
and `-vvv`).

The first level prints only the executed tests, the second level prints
also the test that were skipped either because the module didn't have
tests or because the module didn't matched the patterns in `-p`
arguments. The third level prints also test skipped because a test
failed and `-k` was not used.

The `--summary` option is also adjusted to print the skipped tests only
when the verbose level higher than 1.

Fixes #413.