Detect and handle race condition of CachingDataLoaderBase.load()
With fibers it is possible that a program does reentrant calls of
CachingDataLoaderBase.load() while load() is fetching data for a requested
record from an external source. When a reentrant call uses the same key as
the pending call the reentrant call returns a wrong value.
This commit adds a way of detection this situation, and reentrant calls to
load() return null if it happens.