UnitTestRunner: Fix error printing
After a refactor done in c59d71e12e8dba1dffac1af1c023578d3f24773f,
errors were not printed to Stderr anymore, they were only written to the
XML file (if one was generated). This commit fixes that and prints the
error as before too with an extra enhancement: now the verbose output
and the error output are not mixed.