Allow file redirection of streaming portion
Currently, attempting to redirect the output of an application using
AppStatus to a file from the command line (using > or >>) results in
strange binary characters to be saved in the file. This commit allows
the streaming portion to be saved to the file correctly, while the
static portion is ignored.