Makd: Build integration tests and separate slow tests
This commit automatically builds and runs any program found as
`test/*/main.d` when running `make test` (and `make integrationtest`).
Also, a `fasttest` target is added, which by default only runs the (also
new) `fastunittest` target. This target excludes any module declared in
a file with the suffix `_slowtest.d`.

Please refer to the documentation for the full details.

Fixes #418.