Separate epoll client lists only if timeouts really happened
Currently after epoll_wait() returned all selected keys are looked up in
the list of timed out clients, regardless of whether timeouts really
happened or not. This is inefficient in applications which do not use
timeouts, especially if they have a high number of active connections.
This patch disables the list comparison if no timeouts happened.