Handle selected keys straight away without buffering in selected_set
The selected keys can be handled straight away rather than first separating
the client lists and handling the selected clients afterwards. This is now
done in handleClients() which is a modification of the former
separateClientLists() into which handleTimedOutClients() is merged.
Essentially, now if select() returns, all selected keys and timed out
clients are handled.
This also eliminates the need of EpollSelectDispatcher.selected_keys.