README: Adopt new release notes file
The changelog in the README file is becoming unmanageable and GitHub
provides a nice Releases section where a changelog can be added. Instead
of keep growing the README file with a changelog, now the changelog is
maintained in a temporary file RELEASE_NOTES.md. This files is flushed
when a new release is done and the contents are moved to the release
notes field of the GitHub release. The new file is in markdown format
because GitHub releases don't support reST yet.