app: Add unittest to ArgumentsExt
A unittest is added to the ArgumentsExt class, just to test the new
support to use --help even when required arguments are not passed. To be
able to test this without polluting stdout/stderr, some complexity was
introduced to the class, to be able to override where to write messages
and errors to.