Add config sanity check function to unittest block
Add a function in the unittest block to verify the sanity of a parsed
configuration. This function would simplify checking the results of
other unit tests added in future.

Since this function relies on iterators, it obviates the need for a
separate section in the unittest block dedicated to checking the working
of iterators.