Avoid assigning to ctx.value early on
Currently in the parseLine method, the context variable 'ctx.value' is
assigned early on for no real useful reason. This causes unnecessary
confusion if the line being parsed actually doesn't contain a value at
all (it may be simply a category line). Instead the input variable
'line' itself can be used as long as possible.