Intelligently prune an old configuration
When a new configuration needs to be parsed (either from a file or from
a string) such that it overwrites a previously parsed configuration, it
is currently done in a rather primitive manner. The entire existing
configuration is deleted before parsing the given file or string.

This is not a good approach because in practice, config file changes
across different parses are expected to be small. Most properties in the
config file would actually remain unchanged in different parses, so they
need not even be touched from the parsed configuration.

This commit associates a boolean flag with each property in the
configuration. At the start of every new parse, the flags of all
properties are cleared. If during the parse, the property is found in
the configuration, then only its flag is set. New properties found
during the parse are added to the configuration with their flags set. At
the end of the parse, all properties that have the flag still cleared
are removed. This ensures a correct parse while at the same time
restricting modifications to the configuration to the bare minimum.

Making this change is inconsistent with how configs are overridden in
the ConfigExt module. Hence the parsing logic in this module also needs
to be updated so that config properties can be properly overridden via
the command-line.