Allow parseString to clear an old configuration
The config parser provides two APIs for parsing a configuration:
    * parse       -> parses the configuration from a file
    * parseString -> parses the configuration from the given string

When parsing from a file, it is possible to clear a previously parsed
configuration (albeit in a rather primitive manner) using the second
boolean parameter, but when parsing from a string, clearing a previously
parsed configuration is not possible.

This commit adds the ability to clear a previously parsed configuration
using 'parseString' also.