Config: Trim category names (as with keys and values)
Now the "[category]" in config files will be trimmed, so "[ category ]"
and "[category]" will be parsed both as "category". This is done to
match the behaviour of keys and values, which are trimmed too. For now
a warning is printed if the trimming of the string yielded a different
result, so updating configuration files should be easy.
This warning should be removed in some future.

The ConfigExt from the app framework will also trim categories passed
through the command-line. These are much harder to get wrong, as adding
spaces in the command-line is not usual, so no warnings are printed
here, they are just silently trimmed.

Also, a bunch of tests are added to the ConfigParser to verify how
white spaces are handled.