app: Print help even with parse errors
The ArgumentsExt object now checks for the --help flags before checking
for parser errors. This is particularly useful when specifying "required"
options, because it's considered a parser error if they are not present,
so you can't run ./app --version or ./app --help in this case, otherwise
you'll get an error about the missing arguments instead of the information
you asked for.

This is really a hack, applications should not use required options,
they should use positional arguments for required input (options should
be purely optional).

A better fix will be trying to inspect the results of the parsing and
ignore only required options errors.

Fixes #499.