DigitGrouping: Remove extra space appended
When formatting a number into a string with binary prefixes, it is
possible that the lowest order (i.e. the final part less than 1024 that
has no prefix) simply doesn't exist. This could happen if the given
number is an exact multiple of 1024. For such numbers, an extra space
gets appended (when formatting the previous level). This commit removes
that extra space.