Add delegate-get method to TokyoCabinetM
The old get() method copies the value from the database into the user's
buffer. This is not always required -- for example, in the case where the
record value is simply being written straight to disk, there's no need to
copy it elsewhere in memory. The new function allows both use cases to be
supported.