Pass connection handler info interface to error delegate
This allows the delegate to access information about the connection which
has experienced the error, for example the address and port of the
connected client. This can be very useful for logging.

Note that a github search indicates that this feature is only being used in
swarm (which will be adapted).