BucketSet.clear(): Add documentation usage warning
In some occasions when using hash-maps (which uses BucketSet internally)
and on calling hash-map.clear() method, the hash-map is not properly
cleared and a cyclic method is introduced inside the bucket-set on next
usage of the map.
On the other hand calling hash-map clearErase() method works properly and
doesn't have the mentioned side-effect.

This commit adds a usage warning about the issue inside the methods
documentation for both the clear() and clearErase() methods.