Implement moving average for irregular data
This patch provides a struct implementation of a moving average for
data that is assumed to be updated at irregular intervals.  The values
being averaged are assumed to be floating point (or at least aggregated
as such), while time can be indicated using an arbitrary integral or
floating point type.

The algorithm used is the double-exponential smoothing method described
in section 3 (eqs. 16-22) of Cipra, T. (2006) 'Exponential smoothing for
irregular data', Applications of Mathematics 51 (6): 597-604
<http://dml.cz/handle/10338.dmlcz/134655>.