AppStatus: Add colour support
Support setting the foreground and background colours of both static and
streaming lines of the AppStatus class.

When writing directly to Stdout, the colour can be changed multiple
times in a single line. But this is not possible in case of AppStatus,
wherein a given line can have only one colour. This is because a
streaming line is actually a log event sent to the insert console
appender while formatting and display of static lines happens via
different methods.