ConfigExt: Avoid repeating defaults
Tango automatically shows the default values of an argument when
displaying the help message (if the default values have been set). When
setting up the arguments of the ConfigExt extension, the default values
are configured and also added to the help message resulting in the
default values appearing twice in the help message.
To prevent this, don't manually add the default values to the help
message.