Replace local copy of Makd with a submodule
Now Makd is hosted as a standalone project. This commit adapts ocean to use
the new project, which implies some changes in the generated Version module
to make it more portable across projects (using an associative array to
store the version strings instead of a class).

The old scripts are updated to use the new Version scheme, but are
deprecated and will emit a warning.