StatsLog: Add addObject, which add an object with a name to a category
`addObject` will replace `addSuffix` in a more structured way to allow
one to add an 'instance' (or object) of a 'category' to the log,
which permit finer statistics to be logged.

Also, since there is only tye of input any more - aggregates -, the old
`formatXXX` don't make sense anymore, and we just use `format` now.

Part of sociomantic/ocean#621