D2: Get rid of static const, just use const
'static const' and 'const' have a different meaning in D1 and D2.
In D1, they both equal to manifest constant, while in D2 they are not.
Since the 'const' as in D2 const doesn't exist in D2, d1tod2fix will replace
'const' with 'enum', thus we're cleaning the path for it.