test/unixsocket: Don't clearn using Process("rm -fr {}")
This yielded a small rewrite on the way the `fork()` is done, as all the
`scope (exit)` clauses were also called in the child process, we need to
be more careful.

First, the `fork()` is done before creating the server socket, it doesn't
make sense to have the socket fd of the server in the client too.

Then, a simple retry mechanism was implemented in the client, since it's
forked before the server socket is bound, the connection will fail.

Finally, the `scope(exit)` are placed where they need to go, so when
possible they are only present in the parent process, and when is not
possible, they are disabled in the child process.