StatsLog: Use '/' as a separator - Add a new overload of formatValue
StatsLog: Add a new overload of formatValue

Currently there is no "official" hierarchy scheme.
One project (swarmnodes) is known to use addSuffix to create
its own hierarchy, using '_' (underscore) as a hierarchy
separator. A library solution can't use any character that
would be valid inside an identifier, for obvious collision
concern, which is why we went with slash.

We plan to replace `addSuffix` with a new method, `addObject`.
This requires a new version of `formatValue` which allows an
instance name and category to be specified, along with a value to
be logged.